+++ doc/CHANGELOG.markdown
@@ -5,6 +5,8 @@ Filez-2.0.0 (WIP)
 * Fixed PHP-5.3 incompatibilities
 * Added "show_credit", "help_url" and "bug_report_href" in the looknfeel section of filez.ini
 * Added user documentation
+* Email are now sent under the name of the connected user
+* Removed debug output when application isn't in debug mode
 
 Filez-2.0.0-BETA2
 * Added 'password' field to fz_file table. You need to run 'db.migration-filez-2.0.0-1.sql'